chore: refactor code, custom hooks, type file
