+++ hooks/index.ts
@@ -1,1 +1,2 @@
 export {default as useHandleDestroyAnimated} from './useHandleDestroyAnimated';
+export {default as useMintAndTransfer} from './useMintAndTransfer';
